Split admin-api-listings tests into a folder and drop its max-lines exemption
The admin listings JSON API test file had grown past 1,000 lines and was
grandfathered into biome's `noExcessiveLinesPerFile` exemption list. Split
it into `test/admin-api-listings/` — grouped by what each part checks
(reads, creates, updates, lifecycle actions, and the body-to-input
parsers) — so every file sits well under the limit, then removed the
file's exemption entry.

All 86 original tests carry over unchanged and pass. jscpd stays at 0%
duplication after the split.
