+++ biome.json
@@ -159,7 +159,6 @@
         "src/shared/db/settings.ts",
         "src/ui/templates/fields.ts",
         "src/ui/templates/public/reservations.tsx",
-        "test/admin-api-listings.test.ts",
         "test/e2e/accounting.test.ts",
         "test/e2e/duration-days.test.ts",
         "test/lib/attendee-merge.test.ts",
