+++ test/admin-api-listings/read.test.ts
@@ -1,0 +1,66 @@
+import { expect } from "@std/expect";
+import { describe, it as test } from "@std/testing/bdd";
+import { handleRequest } from "#routes";
+import {
+  apiRequest,
+  assertJson,
+  createTestApiKeyToken,
+  createTestListing,
+  describeWithEnv,
+  mockRequest,
+  requestAsSession,
+  testCookie,
+  testCsrfToken,
+} from "#test-utils";
+
+describeWithEnv("Admin API - Listings", { db: true }, () => {
+  describe("GET /api/admin/listings/:listingId", () => {
+    test("returns single listing by ID", async () => {
+      const listing = await createTestListing({ name: "Detail Listing" });
+      const apiKey = await createTestApiKeyToken();
+
+      await assertJson(
+        apiRequest(`/api/admin/listings/${listing.id}`, { apiKey }),
+        200,
+        (body) => {
+          expect(body.listing.name).toBe("Detail Listing");
+          expect(body.listing.id).toBe(listing.id);
+          expect(body.listing.slug_index).toBeUndefined();
+        },
+      );
+    });
+
+    test("returns 404 for non-existent listing", async () => {
+      await assertJson(apiRequest("/api/admin/listings/99999"), 404, (body) => {
+        expect(body.error).toBe("Listing not found");
+      });
+    });
+
+    test("returns 401 without auth", async () => {
+      const response = await handleRequest(
+        mockRequest("/api/admin/listings/1"),
+      );
+
+      expect(response.status).toBe(401);
+    });
+
+    test("works with cookie+CSRF auth", async () => {
+      const listing = await createTestListing({ name: "Cookie Detail" });
+      const cookie = await testCookie();
+      const csrfToken = await testCsrfToken();
+
+      await assertJson(
+        handleRequest(
+          requestAsSession(`/api/admin/listings/${listing.id}`, {
+            cookie,
+            csrfToken,
+          }),
+        ),
+        200,
+        (body) => {
+          expect(body.listing.name).toBe("Cookie Detail");
+        },
+      );
+    });
+  });
+});
